@@ -13,38 +13,39 @@
     "test": "echo \"No test specified\" && exit 0"
   },
   "dependencies": {
-    "@quasar/extras": "^1.0.0",
-    "@types/leaflet": "^1.9.3",
-    "@types/leaflet-draw": "^1.0.7",
+    "axios": "^1.2.1",
+    "@quasar/extras": "^1.16.4",
+    "quasar": "^2.6.0",
+    "vue": "^3.4.18",
+    "vue-router": "^4.0.12",
     "@vueuse/components": "^10.6.1",
     "@vueuse/core": "^10.1.2",
     "@vueuse/integrations": "^10.7.0",
-    "axios": "^0.21.1",
-    "core-js": "^3.6.5",
     "jwt-decode": "^3.1.2",
     "leaflet": "^1.9.4",
     "leaflet-draw": "^1.0.4",
     "moment": "^2.29.4",
     "qrcode": "^1.5.3",
     "qrcode.vue": "^3.3.3",
-    "quasar": "^2.6.0",
     "swiper": "^9.3.1",
-    "vue": "^3.0.0",
     "vue-currency-input": "^3.0.4",
-    "vue-router": "^4.0.0",
     "vue-timer-hook": "^1.0.60",
     "yandex-metrika-vue3": "^1.2.0"
   },
   "devDependencies": {
-    "@quasar/app-webpack": "^3.0.0",
-    "@types/lodash": "^4.14.191",
-    "@types/node": "^12.20.21",
-    "@typescript-eslint/eslint-plugin": "^5.10.0",
-    "@typescript-eslint/parser": "^5.10.0",
-    "eslint": "^8.10.0",
-    "eslint-config-prettier": "^8.1.0",
+    "@typescript-eslint/eslint-plugin": "^6.6.0",
+    "@typescript-eslint/parser": "^6.6.0",
+    "eslint": "^8.11.0",
     "eslint-plugin-vue": "^9.0.0",
-    "prettier": "^2.5.1"
+    "eslint-config-prettier": "^9.0.0",
+    "prettier": "^3.0.3",
+    "@types/node": "^20.5.9",
+    "@quasar/app-vite": "^2.0.0-beta.3",
+    "autoprefixer": "^10.4.2",
+    "typescript": "^5.2.2",
+    "@types/leaflet": "^1.9.3",
+    "@types/leaflet-draw": "^1.0.7",
+    "@types/lodash": "^4.14.191"
   },
   "browserslist": [
     "last 10 Chrome versions",
@@ -58,7 +59,7 @@
     "last 5 Opera versions"
   ],
   "engines": {
-    "node": ">= 12.22.1",
+    "node": "^24 || ^22 || ^20 || ^18",
     "npm": ">= 6.13.4",
     "yarn": ">= 1.21.1"
   }
