@@ -1,8 +1,10 @@
 /* eslint-disable */
 
+/// <reference types="vite/client" />
+
 // Mocks all files ending in `.vue` showing them as plain Vue instances
 declare module '*.vue' {
-  import type { DefineComponent } from 'vue';
-  const component: DefineComponent<{}, {}, any>;
-  export default component;
+  import type { DefineComponent } from 'vue'
+  const component: DefineComponent<{}, {}, any>
+  export default component
 }
